ExecutionService の Start/Cancel を ExecutionLifecycleCommandService に分離し、ライフサイクルコマンドの責務を独立させる。
